feat: working with customer profile
