+++ src/Bookings.js
@@ -26,7 +26,6 @@ const Bookings = () => {
       });
   }, []);
 
-  console.log({bookings})
   return (
     <div className="App-content">
       <div className="container">
