+++ src/CustomerProfile.js
@@ -2,14 +2,16 @@
 
 function CustomerProfile (props) {
     
-    function profileHandler () {
-        console.log('Customer Profile', props.id)
-
-        // <p> Customer Profile </p>
-    }
-
     return (
-        <button onClick = {profileHandler}>Show Profile</button>
+        <div class="card">
+        <div class="card-body">
+          <h5 class="card-title">Customer Profile</h5>
+          <p class="card-text"> Customer ID: </p>
+          <p class="card-text"> Customer Email: </p>
+          <p class="card-text"> Customer Status: </p>
+          <p class="card-text"> Customer Phone: </p>
+        </div>
+      </div>
     )
 }
